server/requirements.txt
python-decouple==3.8
python-magic==0.4.27
Pillow==10.3.0
PyMuPDF==1.24.2
ujson
hashids
pyjwt
